Refactorized code around Tux module: - Fixed several sign issues - Use STL containers - Simplified code
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@320 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ audio.h
@@ -106,7 +106,7 @@ public:
 	bool Play (const string& name, int loop);
 	bool Play (size_t musid, int loop, int volume);
 	bool Play (const string& name, int loop, int volume);
-	bool PlayTheme (size_t theme, int situation);
+	bool PlayTheme (size_t theme, size_t situation);
 	void Refresh (const string& name);
 	void Halt ();
 	void FreeMusics ();
